Add files via upload
Cameras are complex pieces of hardware that are able to capture 2D images of 3D objects. Indeed, to model real cameras precisely, one need to estimate many parameters including lens distortion, ISO, focal length, and exposure time. In the following, we restrict our focus to the projective camera model.

![](https://storage.googleapis.com/tensorflow-graphics/notebooks/intrinsics/camera_model.jpg)

 This model is composed of two parameters that are often referred to as intrinsic parameters:
- the principal point, which is the projection of the optical center on the image. Ideally, the principal point is close to the center of the image.
- the focal length, which is the distance between the optical center and the image plane. This parameters allows to control the level of zoom.

This notebook illustrates how to use [Tensorflow Graphics](https://github.com/tensorflow/graphics) to estimate the intrinsic parameters of a projective camera. Recovering these parameters is particularly important to perform several tasks, including 3D reconstruction.

In this Colab, the goal is to recover the intrinsic parameters of a camera given an observation and correspondences between the observation and the render of the current solution.  Things are kept simple by only inserting a rectangle in the 3D scene, and using it as the source of correspondences during the optimization. The minimization is performed using the Levenberg-Marquardt algorithm.